Manage to handle zoom.
